requirements.txt
altair==6.0.0
astroid==4.0.2
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
debugpy==1.8.18
decorator==5.2.1
dill==0.4.0
executing==2.2.1
fastjsonschema==2.21.2
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
isort==7.0.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mccabe==0.7.0
narwhals==2.13.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pydeck==0.9.1
Pygments==2.19.2
pylint==4.0.4
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.52.1
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
watchdog==6.0.0
wcwidth==0.2.14
